factor out root command
